LayoutTests/fast/css/content-language-late.html

<!DOCTYPE html>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p>Test for <a href="https://bugs.webkit.org/show_bug.cgi?id=76701">bug 76701</a>:
map HTTP-EQUIV content-language to -webkit-locale. This particular test tests
that a content-language is used even if it appears at the end of the document.
This expectation may change, see bug.  According to the HTML 5 spec, a meta
element should be processed when it is inserted into the document, regardless
of where. IE and Firefox also seem to process the content-language at the end
of the document.
</p>
<div id="console"></div>
<div id="x"></div>
<div id="y" lang="ar"></div>
<script>
function languageOfNode(id) {
    var element = document.getElementById(id);
    return window.getComputedStyle(element).webkitLocale;
}
shouldBeEqualToString("languageOfNode('x')", "ja-JP");
shouldBeEqualToString("languageOfNode('y')", "ar");
</script>
<script src="../js/resources/js-test-post.js"></script>
<meta http-equiv="content-language" content="ja-JP">
</body>
</html>
