third_party/blink/web_tests/fast/css/content-language-late-expected.txt

Test for bug 76701: map HTTP-EQUIV content-language to -webkit-locale. This particular test tests that a content-language is used even if it appears at the end of the document. This expectation may change, see bug. According to the HTML 5 spec, a meta element should be processed when it is inserted into the document, regardless of where. IE and Firefox also seem to process the content-language at the end of the document.

FAIL languageOfNode('x') should be "ja-JP". Was auto.
PASS languageOfNode('y') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
